# perf(buffer)!: apply SSO technique to text buffer in `buffer::Cell`