@rhysd 2023-10-28
I noticed performance regression that a benchmark with 100 tick iterations got ~10% slower.

After some investigation with CPU sampling profiler, I found that calling `clear()` then pushing new string/char is slightly slower with `CompactString` (I guess it checks its length to determine to use inline storage or spill to heap memory).

I fixed it at c5ecda6 and confirmed the fix with hyperfine:

```console
$ hyperfine ./demo_before ./demo_after
Benchmark 1: ./demo_before
  Time (mean ± σ):      95.2 ms ±   2.0 ms    [User: 92.1 ms, System: 1.8 ms]
  Range (min … max):    92.7 ms … 101.2 ms    30 runs

Benchmark 2: ./demo_after
  Time (mean ± σ):      75.3 ms ±   0.9 ms    [User: 72.5 ms, System: 1.6 ms]
  Range (min … max):    73.6 ms …  77.9 ms    38 runs

Summary
  ./demo_after ran
    1.26 ± 0.03 times faster than ./demo_before
```