@a-kenji 2023-10-29
I can reproduce the improvements on:
[tui-term-main](https://github.com/a-kenji/tui-term/tree/main) vs [tui-term-sso](https://github.com/a-kenji/tui-term/tree/upstream/sso):

![Pasted image 1](https://github.com/ratatui-org/ratatui/assets/65275785/31005c64-6cf9-4102-ba9a-535b94b40f66)
Main:
![main_divan](https://github.com/ratatui-org/ratatui/assets/65275785/3decbcec-49b3-482c-bc6c-684074149ad4)
Sso:
![sso_divan](https://github.com/ratatui-org/ratatui/assets/65275785/2a968a94-9a63-4d94-8f2d-c591c593cad0)

Note: The benchmarks are all done with a grid of `80:24`, so I am likely looking at a minimum of performance gained. Also even tough I use `buffer::Cell` directly, I did not need to change any code.