@joshka 2023-10-30
Thanks for doing tests on this - this was missing from the previous approaches.

Some questions:

1. How does this compare to the time spent rendering in the actual terminal?
2. I'd suggest using something like https://github.com/ratatui-org/ratatui/pull/583 as a test for this (it changes every cell and the background / foreground every frame, so in theory will ensure that we see the effect on the worst performing app.
3. We could perhaps simplify to create a random full buffer every frame instead.
4. It would be a good idea to tests ~10K cells (200x50) as this tends to be the size of a 16in Macbook with some default font sizes (as well as 80x24 for a min and 255x255 for a max)
5. Is it worth testing the other string types?
6. What are your thoughts on the absolute magnitude of the change (the relative 14% change obscures that this is 0.1ms - what sorts of scenarios is that worth worrying about in?
7. Would it be worth making the Cell fields private to avoid exposing this internal detail?

Some previous similar work (tagging for their insights / interest):
- https://github.com/ratatui-org/ratatui/issues/356 by @jamespharaoh
- https://github.com/ratatui-org/ratatui/pull/231 by @Philipp-M
- https://github.com/fdehau/tui-rs/pull/644 by @rhysd :)