@rhysd 2023-10-30
@a-kenji Thank you for your trial. It is very helpful to try this branch by practical external widgets. I'm glad to hear that benchmarks on your side were improved.

@joshka Thank you for the feedback.

> How does this compare to the time spent rendering in the actual terminal?

No backend is modified. So the difference is caused only by accessing to the content (e.g. [crossterm](https://github.com/ratatui-org/ratatui/blob/9f371000968044e09545d66068c4ed4ea4b35d8a/src/backend/crossterm.rs#L164)).

Theoretically it only adds one branch per accessing one cell, since `CompactString` needs to check if it's using inline storage or heap buffer. The branch only adds a few instructions and we can ignore the impact.

Practically, I'll try to measure the actual impact with real backend. I'll post the result later.

> I'd suggest using something like https://github.com/ratatui-org/ratatui/pull/583 as a test for this (it changes every cell and the background / foreground every frame, so in theory will ensure that we see the effect on the worst performing app.

OK, I'll measure performance with `colors_rgb` and `TestBackend` later.

> We could perhaps simplify to create a random full buffer every frame instead.

Do you mean measuring creating `Buffer` instance alone? It's possible but I feel initial rendering measurement is sufficient for this.

> It would be a good idea to tests ~10K cells (200x50) as this tends to be the size of a 16in Macbook with some default font sizes (as well as 80x24 for a min and 255x255 for a max)

OK, I'll try 200x50 and 255x255. 80x24 was already measured by @a-kenji's benchmark and I feel that's sufficient.

> Is it worth testing the other string types?

I considered the following crates:

- `smol_str` : MSRV policy does not fit to this crate. And the author said that [`ecow` is the better implementation](https://www.reddit.com/r/rust/comments/117ksvr/comment/j9ehr5t/?utm_source=reddit&utm_medium=web2x&context=3).
- [`ecow`](https://github.com/typst/ecow) : This is a good alternative. However, in terms of the following points, I selected `compact_str`
  - Its inline storage size is smaller than `compact_str`
  - It internally manages its buffer with reference counter for O(1) clone. This is effective when cloning string happens frequently but it is not the case of this crate's usage. And increasing/decreasing atomic reference counter has additional cost.

If you have other alternative, I'll take a look.

> What are your thoughts on the absolute magnitude of the change (the relative 14% change obscures that this is 0.1ms - what sorts of scenarios is that worth worrying about in?

I'm not sure the fact that rendering on each tick is 1.14x faster has real impact on user's use case. Rendering happens per some intervals (e.g. 100ms) so program has enough budget for rendering each frames with current main branch. CPU usage of program may slightly go down tough.

The most important point in regards to performance is initial rendering since a program usually tries to render the first paint and it is CPU intensive. The absolute values of my benchmark with only initial rendering are 5.0ms (sso) vs 8.8ms (main). I feel 3.8ms is large enough to be considered.

> Would it be worth making the Cell fields private to avoid exposing this internal detail?

Well, that depends on maintainers. I have no strong opinion. `CompactString`'s API is almost compatible with `String`. Users of `buffer::Cell` (I guess there would be only a few direct users though) can use the field almost the same as `String`. `.into()` method automatically converts string types into `CompactString` too.

In general, exposing struct fields can cause an issue like this. If I were a maintainer of this crate, I would add a builder for `Cell` and add accessor methods instead of exposing fields.